+++ docs/additional_examples_and_pseudo_code/active_symbols.py
@@ -1,4 +1,0 @@
-from tekhsi import TekHSIConnect
-
-with TekHSIConnect("192.168.0.1:5000") as connection:
-    print(connection.activesymbols)