+++ CHANGELOG.md
@@ -18,6 +18,10 @@ Valid subsections within a version are:
 
 Things to be included in the next release go here.
 
+### Changed
+
+- Updated the documentation by moving portions from the Basic Usage page to the API docs.
+
 ---
 
 ## v1.0.0 (2024-09-20)